economy with SQLITE3/requirements.txt
python-dotenv>=0.21.0
